delivary/requirements.txt
Django==4.1.6
django-extensions==3.2.1
django-softdelete==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1